Prefer f64 over f32 in mandelbrot calculations
This will allow more precision while we go deeper into the mandelbrot
set. Would eventually like to try Ratios of BigInts over f64, but saving
that for farther down the road.

+++ lib/mandelbrot/src/image_create.rs
@@ -27,7 +27,7 @@ pub fn write_png(prefix: &str, frame: Frame) -> Result<(), Box<error::Error>> {
     return Ok(());
 }
 
-fn gen_pixel(divergence: f32) -> image::Rgba<u8> {
+fn gen_pixel(divergence: f64) -> image::Rgba<u8> {
     let div_u8 = divergence as u8;
     if div_u8 == 0 {
         return image::Rgba([0, 0, 0, 255]);
@@ -39,19 +39,19 @@ fn gen_pixel(divergence: f32) -> image::Rgba<u8> {
 pub struct Frame {
     screen_width: u32,
     screen_height: u32,
-    min_x: f32,
-    min_y: f32,
-    plot_width: f32,
-    plot_height: f32,
+    min_x: f64,
+    min_y: f64,
+    plot_width: f64,
+    plot_height: f64,
 }
 
 impl Frame {
     pub fn new(screen_width: u32,
                screen_height: u32,
-               min_x: f32,
-               max_x: f32,
-               min_y: f32,
-               max_y: f32)
+               min_x: f64,
+               max_x: f64,
+               min_y: f64,
+               max_y: f64)
                -> Frame {
         let plot_width = max_x - min_x;
         let plot_height = max_y - min_y;
@@ -65,9 +65,9 @@ impl Frame {
         };
     }
 
-    fn get_coord_for_pixel(&self, screen_x: u32, screen_y: u32) -> (f32, f32) {
-        let x_percent = f32::from(screen_x as u16) / f32::from(self.screen_width as u16);
-        let y_percent = f32::from(screen_y as u16) / f32::from(self.screen_height as u16);
+    fn get_coord_for_pixel(&self, screen_x: u32, screen_y: u32) -> (f64, f64) {
+        let x_percent = f64::from(screen_x) / f64::from(self.screen_width);
+        let y_percent = f64::from(screen_y) / f64::from(self.screen_height);
 
         let x = self.plot_width * x_percent + self.min_x;
         let y = self.plot_height * y_percent + self.min_y;
